requirements.txt
fastapi==0.95.1
openai==0.27.4
pydantic==1.10.7
python-dotenv==1.0.0
uvicorn==0.21.1
twilio==8.2.0